[TD-240] 가게의 이미지 기능을 별도의 컨트롤러 및 서비스로 분리해서 관리
